More moving code about to make stuff easier to grasp
